moved tests and change packages and imports
