+++ Model/src/main/scala/model/gameComponent/gameBaseImpl/Color.scala
@@ -1,4 +1,4 @@
-package de.htwg.se.uno.model.gameComponent.gameBaseImpl
+package model.gameComponent.gameBaseImpl
 
 enum Color:
   case Red, Yellow, Green, Blue, Special
