+++ Model/src/main/scala/model/gameComponent/gameBaseImpl/Value.scala
@@ -1,4 +1,4 @@
-package de.htwg.se.uno.model.gameComponent.gameBaseImpl
+package model.gameComponent.gameBaseImpl
 
 enum Value:
   case Zero, One, Two, Three, Four, Five, Six, Seven, Eight, Nine, DirectionChange, Suspend, PlusTwo, PlusFour, ColorChange
