+++ Model/src/test/scala/model/CardSpec.scala
@@ -1,6 +1,6 @@
-package de.htwg.se.uno.model
+package model
 
-import de.htwg.se.uno.model.gameComponent.gameBaseImpl.{Card, Color, Value}
+import model.gameComponent.gameBaseImpl.{Card, Color, Value}
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
