+++ Tools/src/main/scala/tools/util/UndoManager.scala
@@ -1,4 +1,4 @@
-package de.htwg.se.uno.util
+package tools.util
 
 import scala.util.Try
