+++ src/test/scala/de/htwg/se/uno/aview/gui/CardPanelSpec.scala
@@ -1,7 +1,7 @@
 package de.htwg.se.uno.aview.gui
 
 import de.htwg.se.uno.controller.controllerComponent.controllerBaseImpl.Controller
-import de.htwg.se.uno.model.gameComponent.gameBaseImpl._
+import model.gameComponent.gameBaseImpl._
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
